Add character movement with path finding
A BarriersComponent keeps a list of points that
the player cannot walk through and the
corresponding RectangleComponents that are drawn
on screen. The points are part of the minigrid, a
2d grid of integers that A* operates on.

We use the a_star_algorithm to calculate a set of
points that define a path that avoids all barriers.

We use an enhanced enum for direction that has
up, down, left, right as well as the diagonals.
The Player is a SpriteAnimationGroupComponent that
maps direction to the appropriate animation.

We now create the FlameGame object once in
main.dart and pass into the app widget to avoid
recreating the object when the widget tree re-renders.

Flame's game and world:
- We make a world component by extending flame's
World class and the world compent adds all other
components that make up the game world.
- Extend FlameGame where we set the world component,
load texture images and setup the camera.

Added a PlayerComponent that keeps the position
of the player and draws the sprite animation. The
PlayerComponent also provides movement either by
adding MoveEffects at the same time as changing
the animation to the relevant walking direction.
The anchor point draws the 32x64 sprite in the
appropriate place that corresonds to the grid point
that matches the position of the component.

Added a GridComponent made up of a set of
one pixel wide RectangleComponents that allow
drawing the grid on the canvas.

Added PathComponent takes the barriers and uses
the a_star_algorithm to calculate a set of points
between the start (player position) and end
(clicked point). The PathComponent also keeps the
set of "large grid points" that correspond to the
bottom left of the large grid squares we see on the
canvas, as well as RectangleComponents that can be
used to draw the path on the canvas.

We calculate a direction for each path segment
from the points generated by a_star_algorithm and
use the directions to add MoveEfftects to the Player.

We round the Player position before calculating
the miniGrid position, as position values are
doubles and do not necessarily hold exact values.