Add multiplayer (#8)
* Add dependencies

- tech_world_networking_types to share interfaces between server and client
- web_socket_channel to connect to multiplayer server

* Move to a service locator for state management

* Move flutterfire auth code into an AuthService

* Rename firebase_auth to auth_gate

To reflect the file name containing the AuthGate widget

* Create NetworkingUser, AuthUser, implementing User

PlayerComponent, AuthUser and NetworkUser all implement
User, which has just a String id and a displayName. This reflects the
relationship between the classes, ie. that they all represent a User, in
different contexts.

* Create a PlayersService

That has relationships with the AuthService, NetworkingService, and the
World object, tieing everything together.

* Connect the World object to services

The PlayersService, which is connected to the NetworkingService,
passes in streams that the World listens to.

* Access the TechWorldGame through the service locator

* Add a PlayerPath class

Created from a PlayerPathMessage, using the relevant types for moving
players.

* Add constants for the server urls