Retrieve-fresh-token (#22)
* Add new firebase functions to the project

The new firebase function is a callable function that retrieves a
LiveKit token and returns it as the result.

The firebase function called saveDoc was previously being used,
triggered on userCreation so we were only retrieving a token when a
user was created and saving it in the Firestore under the uid, then
the client read from the firestore when it was updated.

The old function resulted in a stale token being used if the an
existing user signed in, which was happening sometimes even on anonymous
sign in (probably the user was still signed in when they came back).

With the new callable function we are now retrieving a new token every
time so the token is always fresh. The token creation is pretty fast
so the added complexity of storing it and checking for freshness
is not a useful tradeoff.

* Update .gitignore to ignore .env file

* Add cloud_functions and upgrade firebase packages

* Acccept automatic change to Generated Plugin Registrant

* Accept functios change to firebase.json

* Update VideoTrackRenderer to use new API

After updating the livekit_client the API had changed so these
changes update to the equivalent API calls.

* Import collection package

We were using the collection package without directly importing it.

* Change ConnectPage to retrieve a new token

Disable the Connect and Sign Out buttons until a new token
has been retrieved to ensure the user has a fresh token before
attempting to connect.